상점 아이템 오류 발생 수정
상점에서 아이템을 구매할 때 아이템을 복사해서 넣는 방식이 아닌 생성된 아이템을 가져오는 방식이 되어서 아이템을 사용하여 아이템이 제거 되었을 때 상점 내부의 아이템도 함께 제거되어 발생하는 문제였습니다.
 복사 생성자를 통해 아이템 구매 시 복사된 아이템을 인벤토리에 넣어 아이템 본체에는 영향이 안가도록 하였습니다.
