+++ CH2_project/Background.h
@@ -1,22 +1,14 @@
 #pragma once
 #include <iostream>
+#include "Utility.h"
 
 using namespace std;
 
 
 // 임시용 출력확인용
-void title() {
-	
-	cout << endl << endl << " ⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠠⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠓⠀⠀⠀⠀⠀⠀⠀⠀⠀⡏⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀" << endl <<
-		"⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⡰⠊⠑⠀⠀⡃⡹⡄⠸⡄⠀⢴⠁⡴⡑⡑⣆⠀⠀⠀⠀⣰⠑⠙⠴⡀⠙⡏⠁⠀⠀⠀⢜⠎⠊⡶⠊⠱⣂⠠⡇⢀⢾⠈⠓⡄⢀⡖⠙⠘⡮⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀" << endl <<
-		"⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢧⡁⠀⡀⢼⠊⢉⡧⠀⢳⡨⠆⠀⢯⠁⠉⡁⠀⠀⠀⠀⢧⡀⠀⣸⠁⠨⡇⠀⠀⠀⠀⡹⠄⠀⡽⠀⠀⡇⢐⢇⠀⡧⠀⠐⡇⠸⣂⠀⢀⡗⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀" << endl <<
-		"⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⠉⠀⠈⠉⠁⠁⠀⠀⠉⠀⠀⠈⠉⠉⠀⠀⠀⠀⠀⠀⠉⠉⠀⠀⠈⠁⠀⠀⠀⠀⠈⠁⠀⠉⠀⠀⠁⠀⠁⠀⠁⠀⠀⠁⠀⠈⠉⠁⠉⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀" << endl;
+void title();
+
+void gameInfo();
 
 
-}
-
-void gameInfo() {
-	cout << endl << endl << endl << "				여백의 미";
-}
-
-void interface();
+int menuDraw();
