+++ CH2_project/CH2_project.vcxproj.filters
@@ -33,7 +33,10 @@
     <ClCompile Include="Monster.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
-    <ClCompile Include="CursorMove.cpp">
+    <ClCompile Include="Utility.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="Background.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
   </ItemGroup>
@@ -59,7 +62,7 @@
     <ClInclude Include="Background.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
-    <ClInclude Include="CursorMove.h">
+    <ClInclude Include="Utility.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
