+++ CH2_project/Character.h
@@ -3,6 +3,7 @@
 #include <string>
 #include <vector>
 #include <memory>
+#include <conio.h>
 #include "Item.h"
 #include "Shop.h"
