--- CH2_project/CursorMove.h
@@ -1,14 +1,0 @@
-#pragma once
-#include <Windows.h>
-#include <iostream>
-#include <conio.h>
-
-using namespace std;
-
-int keyControl();
-
-void gotoxy(int x, int y);
-
-void eraseCursor();
-
-int menuDraw();
