+++ CH2_project/Item.h
@@ -16,10 +16,11 @@ protected:
 public:
 	Item(const string& name, int price) : name(name), price(price) {} // 생성자
     virtual ~Item() {} // 가상 소멸자
-
+	Item(const Item& other) : name(other.name), price(other.price) {} // 복사생성자
+	virtual Item* clone() const = 0; // 복제용 순수 가상 함수
 	virtual string getName() const;  // 아이템 이름 반환
 	virtual void use(Character* character) = 0; // 캐릭터에게 아이템 사용 (순수 가상함수)
-	virtual int getPrice() = 0; // 아이템 가격 반환
+	virtual int getPrice() const; // 아이템 가격 반환
 	
 };
 
@@ -31,8 +32,9 @@ private:
 
 public:
 	MorPhinePotion() : Item("Morphine", 10), healthRestore(25){} // 초기화 -> 가격 10 회복량 25회복 설정
+	MorPhinePotion(const MorPhinePotion& other) : Item(other), healthRestore(other.healthRestore) {}
+	Item* clone() const override;
 	void use(Character* character) override;
-	int getPrice();
 };
 
 class CocainePotion : public Item {
@@ -41,8 +43,9 @@ private:
 
 public:
 	CocainePotion() : Item("CocainePotion", 20), attackPower(10){} // 초기화 -> 가격 20 공격력 10증가 설정
+	CocainePotion(const CocainePotion& other) : Item(other), attackPower(other.attackPower) {}
+	Item* clone() const override;
 	void use(Character* character) override;
-	int getPrice();
 };
 
 class HeroinPotion : public Item {
@@ -51,8 +54,9 @@ private:
 
 public:
 	HeroinPotion() : Item("HeroinPotion", 20), healthRestore(50){} // 초기화 -> 가격 20 회복량 50회복 설정
+	HeroinPotion(const HeroinPotion& other) : Item(other), healthRestore(other.healthRestore) {}
+	Item* clone() const override;
 	void use(Character* character) override;
-	int getPrice();
 };
 
 class LSDPotion : public Item {
@@ -61,8 +65,9 @@ private:
 	int attackPower; // 공격력
 public:
 	LSDPotion() : Item("LSDPotion", 40), healthRestore(50), attackPower(10){} // 초기화 -> 가격 40, 회복량 50, 공격력 10 증가
+	LSDPotion(const LSDPotion& other) : Item(other), healthRestore(other.healthRestore), attackPower(other.attackPower) {}
+	Item* clone() const override;
 	void use(Character* character) override;
-	int getPrice();
 };
 
 
@@ -72,8 +77,9 @@ private:
 
 public:
 	IVbeltWeapon() : Item("IVbeltWeapon", 0), attackPower(10){} // 초기화 -> 가격0, 공격력 10증가
+	IVbeltWeapon(const IVbeltWeapon& other) : Item(other), attackPower(other.attackPower) {}
+	Item* clone() const override;
 	void use(Character* character) override;
-	int getPrice();
 };
